+++ db/minion.go
@@ -5,10 +5,9 @@ package db
 type Minion struct {
 	ID int `json:"-"`
 
-	Self                bool              `json:"-"`
-	Blueprint           string            `json:"-" rowStringer:"omit"`
-	AuthorizedKeys      string            `json:"-" rowStringer:"omit"`
-	MinionIPToPublicKey map[string]string `json:"-" rowStringer:"omit"`
+	Self           bool   `json:"-"`
+	Blueprint      string `json:"-" rowStringer:"omit"`
+	AuthorizedKeys string `json:"-" rowStringer:"omit"`
 
 	// Below fields are included in the JSON encoding.
 	Role        Role