+++ CHANGELOG.md
@@ -10,6 +10,7 @@ environment variables through the constructor call or by directly accessing the
 .env attribute of the Container instance.
 - Remove the deprecated LoadBalancer.hostname() method in favor of
 LoadBalancer.getHostname().
+- Replace the custom Kelda scheduler for managing containers with Kubernetes.
 
 Release 0.10.0
 -------------